Fix another build break by moving task related code into a general header.
