+++ lib/Dialect/TritonGPU/Transforms/WSDataPartition.cpp
@@ -19,7 +19,6 @@
 #include "triton/Dialect/TritonGPU/IR/Dialect.h"
 #include "triton/Dialect/TritonGPU/Transforms/Passes.h"
 #include "triton/Dialect/TritonGPU/Transforms/Utility.h"
-#include "triton/Dialect/TritonNvidiaGPU/Transforms/Utility.h"
 
 namespace tt = mlir::triton;
 namespace ttg = mlir::triton::gpu;
